// apps/searchneu/app/graduate/common/types.ts
export enum NUPathEnum {
    ND = "Natural/Designed World",
    EI = "Creative Express/Innov",
    IC = "Interpreting Culture",
    FQ = "Formal/Quant Reasoning",
    SI = "Societies/Institutions",
    AD = "Analyzing/Using Data",
    DD = "Difference/Diversity",
    ER = "Ethical Reasoning",
    WF = "1st Yr Writing",
    WD = "Adv Writ Dscpl",
    WI = "Writing Intensive",
    EX = "Integration Experience",
    CE = "Capstone Experience",
  }
// ...
  export type INEUReq = INEUAndReq | INEUOrReq | INEUReqCourse;
  /**
   * A SearchNEU AND prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites that must be completed for this prereq. to be
   *   marked as done.
   */
  export interface INEUAndReq {
    type: "and";
    values: INEUReq[];
  }
  
  /**
   * A SearchNEU OR prerequisite object.
   *
   * @param type   The type of the SearchNEU prerequisite.
   * @param values The prerequisites of which one must be completed for this
   *   prerequisite to be marked as done.
   */
  export interface INEUOrReq {
    type: "or";
    values: INEUReq[];
  }
// ...
  export interface ScheduleCourse2<T> {
    name: string;
    classId: string;
    subject: string;
    prereqs?: INEUAndReq | INEUOrReq;
    coreqs?: INEUAndReq | INEUOrReq;
    nupaths?: NUPathEnum[];
    numCreditsMin: number;
    numCreditsMax: number;
    id: T;
    generic?: boolean;
  }